setupstatedata.py and runallstates.py now look at $REDISTRICTER_BIN and $REDISTRICTER_DATA Also all switched over to expecting data/XX/config/* config modes for congress and state legislatures. Switched runallstates.py to optparse from custom code. Turn off block labelling in mask render, which is failing on one of my JVMs. More verbosity in java task.

+++ generaterunconfigs.py
@@ -8,26 +8,39 @@ import csv
 import os
 import sys
 
+legpath_ = os.path.join(os.path.dirname(__file__), 'legislatures.csv')
+
 from states import codeForState
 
-congress_total = 0
+def run(datadir='data', stulist=None, dryrun=False, newerthan=None):
+	congress_total = 0
+	f = open(legpath_, 'rb')
+	fc = csv.reader(f)
+	for row in fc:
+		stu = codeForState(row[0])
+		if (stulist is not None) and (stu not in stulist):
+			continue
+		configdir = os.path.join(datadir, stu, 'config')
+		if not os.path.isdir(configdir):
+			os.makedirs(configdir)
+		name_part = row[1].split()[0]
+		if name_part == 'Congress':
+			congress_total += int(row[2])
+		outname = os.path.join(configdir, name_part)
+		if newerthan and not newerthan(legpath_, outname):
+			if dryrun:
+				print '"%s" up to date' % (outname)
+			continue
+		if dryrun:
+			print 'would write "%s"' % (outname)
+		else:
+			out = open(outname, 'w')
+			out.write("""datadir: $DATA/%s\ncommon: -d %s\n""" % (stu, row[2]))
+			if (row[2] == '1') or (row[2] == 1):
+				out.write('disabled\n')
+			out.close()
+	return congress_total
 
-f = open('legislatures.csv', 'rb')
-fc = csv.reader(f)
-for row in fc:
-	stu = codeForState(row[0])
-	configdir = os.path.join('data', stu, 'config')
-	if not os.path.isdir(configdir):
-		os.makedirs(configdir)
-	name_part = row[1].split()[0]
-	if name_part == 'Congress':
-		congress_total += int(row[2])
-	outname = os.path.join(configdir, name_part)
-	print (outname, row[2])
-	out = open(outname, 'w')
-	out.write("""datadir: $DATA/%s\ncommon: -d %s\n""" % (stu, row[2]))
-	if (row[2] == '1') or (row[2] == 1):
-		out.write('disabled\n')
-	out.close()
-
-print 'congress_total == %s, should be 435' % congress_total
+if __name__ == '__main__':
+	congress_total = run()
+	print 'congress_total == %s, should be 435' % congress_total
